Final result packet modified

+++ BattleshipNetworkLib/BGNL/Network.h
@@ -79,7 +79,7 @@ public:
 	struct FinalResultPacket
 	{
 		int winCount;
-		int avgTurns;
+		float avgTurns;
 	};
